+++ CHANGES.md
@@ -5,6 +5,7 @@ Notable changes between versions.
 ## Latest
 
 * Fix CoreDNS AntiAffinity spec to prefer spreading replicas
+* Disable Kubelet read-only port ([#324](https://github.com/poseidon/typhoon/pull/324))
 
 #### AWS